fixed user after free when there are multiple elements with equal key
The order of elements with equal keys in the fibonacci heap and the
unordered mutlimap are not always the same. pop() assumed that this is
the case. It could happen that extrac_min() returned a different
FibNode than fstore.find(). The fstore then contains a pointer to an
already deleted FibNode which leads to the use after free problem.